Removed redundant file size calculation and changed file size limit

+++ src/global/global.h
@@ -11,7 +11,7 @@ struct node{
 };
 
 extern int length;
-extern unsigned long long fileSize;
+extern int fileSize;
 extern FILE *input;
 extern FILE *output;
